new oop structure completed
