+++ views/pages/service_status.blade.php
@@ -1,0 +1,39 @@
+<pre id="sambaLog"></pre>
+
+<script>
+    // Control Samba4.service Status == Tab 3 ==
+
+    function tab3(){
+        var form = new FormData();
+        request(API('return_samba_service_status'), form, function(response) {
+            message = JSON.parse(response)["message"];
+            if(message == true){
+                isActiveButton = '<button type="button" class="btn btn-success" disabled>Samba Servisi Aktif !</button>' ;
+                $('#tab3').html(isActiveButton);
+
+                var d1 = document.getElementById('tab3');
+                d1.insertAdjacentHTML('beforeend', '<pre id="sambaLog">   </pre>');
+                sambaLog();
+            } else{
+                isActiveButton = '<button type="button" class="btn btn-danger" disabled>Samba Servisi Aktif Değil !</button>' ;
+                $('#tab3').html(isActiveButton);
+
+            }
+        }, function(error) {
+            showSwal(error.message, 'error', 3000);
+            console.log(error);
+        });
+    }
+
+    function sambaLog(){
+        var form = new FormData();
+        request(API('return_samba_service_log'), form, function(response) {
+            message = JSON.parse(response)["message"];
+            $('#sambaLog').html(message);
+        }, function(error) {
+            showSwal(error.message, 'error', 3000);
+            console.log(error);
+        });
+    }
+
+</script>
