+++ views/pages/trust.blade.php
@@ -1,0 +1,173 @@
+@component('modal-component',[
+        "id" => "createTrustRelationModal",
+        "title" => "Create Trust Relation",
+        "footer" => [
+            "text" => "Create",
+            "class" => "btn-success",
+            "onclick" => "createTrustRelation()"
+        ]
+    ])
+    @include('inputs', [
+        "inputs" => [
+            "Domain Name" => "newDomainName:text:deneme.lab",
+            "IP Address" => "newIpAddr:text",
+            "Type:newType" => [
+                "forest" => "forest",
+                "external" => "external"
+            ],
+            "Direction:newDirection" => [
+                "incoming" => "incoming",
+                "outgoing" => "outgoing",
+                "both" => "both"
+            ],
+            "Create Location:newCreateLocation" => [
+                "local" => "local",
+                "both" => "both"
+            ],
+            "Username" => "newUsername:text",
+            "Password" => "password:password"
+        ]
+    ])
+@endcomponent
+
+@component('modal-component',[
+        "id" => "trustedServerDetailsModal",
+        "title" => "Details",
+        "footer" => [
+            "text" => "Close",
+            "class" => "btn-success",
+            "onclick" => "closeTrustedServerDetailsModal()"
+        ]
+    ])
+@endcomponent
+
+@component('modal-component',[
+        "id" => "deleteTrustedServerModal",
+        "title" => "Warning",
+        "footer" => [
+            "text" => "Cancel",
+            "class" => "btn-success",
+            "onclick" => "closeTrustedServerDetailsModal()"
+        ],
+        "footer" => [
+            "text" => "Delete",
+            "class" => "btn-danger",
+            "onclick" => "destroyTrustRelation()"
+        ]
+    ])
+        @include('inputs', [
+        "inputs" => [
+            "Password" => "password:password"
+        ]
+    ])
+@endcomponent
+
+<button class="btn btn-success mb-2" id="createButton" onclick="showCreateTrustRelationModal()" type="button">Create</button>    
+<div id="trustedServers"></div>
+
+<script>
+
+    var domainName = "";
+    var passwd = "";
+    /**
+     * Showing the servers which have trusted relation with this server
+    */
+
+    function trustedServers(){
+        var form = new FormData();
+        request(API('trusted_servers'), form, function(response) {
+            $('#trustedServers').html(response).find('table').DataTable({
+            bFilter: true,
+            "language" : {
+                url : "/turkce.json"
+            }
+          });;
+        }, function(error) {
+            $('#trustedServers').html("Hata oluştu");
+        });
+        
+    }
+
+    function showTrustedServerDetailsModal(line){
+        var name = line.querySelector("#name").innerHTML;
+        var type = line.querySelector("#type").innerHTML;
+        var transitive = line.querySelector("#transitive").innerHTML;
+        var direction = line.querySelector("#direction").innerHTML;
+        console.log(name);
+        if(name)
+            $('#trustedServerDetailsModal h4.modal-title').html("Details");
+        $('#trustedServerDetailsModal').find('.modal-body').html(
+            "Name".bold() + "</br>" + name + "</br>" + "</br>" +
+            "Type".bold() + "</br>" + type + "</br>" + "</br>" +
+            "Transitive".bold() + "</br>" + transitive + "</br>" + "</br>" +
+            "Direction".bold() + "</br>" + direction + "</br>" + "</br>"
+        );
+        $('#trustedServerDetailsModal').modal("show");
+    }
+
+    function closeTrustedServerDetailsModal(){
+        $('#trustedServerDetailsModal').modal("hide");
+    }
+
+    /**
+     * Deleting the servers which have trusted relation with this server
+    */
+
+    function closeDeleteTrustedServerModal(){
+        $('#deleteTrustedServerModal').modal("hide");
+    }
+
+    function showDeleteTrustedServerModal(line){
+        let name = line.querySelector("#name").innerHTML;
+        domainName = name;
+        $('#deleteTrustedServerModal').find('.modal-body').prepend(
+            "If you destroy trust relation with \"".bold() + name.bold() + "\", please fill the password field.".bold() + "<br><br>");
+        $('#deleteTrustedServerModal').find('.modal-footer')
+            .append('<button type="button" class="btn btn-primary" onClick="closeDeleteTrustedServerModal()">Cancel</button>');
+        $('#deleteTrustedServerModal').modal("show");
+    }
+
+    function destroyTrustRelation(){
+        var form = new FormData();
+        form.append("name", domainName);
+        passwd = $('#deleteTrustedServerModal').find('input[name=password]').val();
+        form.append("password", passwd);
+        closeDeleteTrustedServerModal();
+        trustedServers();
+        request(API('destroy_trust_relation'), form, function(response) {
+            message = JSON.parse(response)["message"];
+            showSwal(message, 'success', 3000);
+        }, function(error) {
+            showSwal(error.message, 'error', 3000);
+        });
+    }
+
+    /**
+     * Creating trust relation with a new server
+    */
+
+    function showCreateTrustRelationModal(){
+        $('#createTrustRelationModal').modal("show");
+    }
+
+    function createTrustRelation(){
+        var form = new FormData();
+        form.append("newDomainName", $('#createTrustRelationModal').find('input[name=newDomainName]').val());
+        form.append("newIpAddr", $('#createTrustRelationModal').find('input[name=newIpAddr]').val());
+        form.append("newType", $('#createTrustRelationModal').find('select[name=newType]').val());
+        form.append("newDirection", $('#createTrustRelationModal').find('select[name=newDirection]').val());
+        form.append("newCreateLocation", $('#createTrustRelationModal').find('select[name=newCreateLocation]').val());
+        form.append("newUsername", $('#createTrustRelationModal').find('input[name=newUsername]').val());
+        form.append("password", $('#createTrustRelationModal').find('input[name=password]').val());
+
+        $('#createTrustRelationModal').modal("hide");
+        request(API('create_trust_relation'), form, function(response) {
+            message = JSON.parse(response)["message"];
+            showSwal(message, 'success', 10000);
+            trustedServers();
+        }, function(error) {
+            showSwal(error.message, 'error', 3000);
+        });
+    }
+
+</script>
